develop menu system.

+++ src/BasicEnemy.java
@@ -14,7 +14,7 @@ public class BasicEnemy extends GameObject {
     }
 
     public Rectangle getBounds(){
-        return new Rectangle(x,y,16,16);
+        return new Rectangle((int)x,(int)y,16,16);
     }
 
     @Override
@@ -24,13 +24,13 @@ public class BasicEnemy extends GameObject {
 
        if(y <= 0 || y >= Game.HEIGHT - 32) velY *= -1;
        if(x <=0 || x >= Game.WIDTH - 16) velX *= -1;
-       handler.addObject(new Trail(x,y,ID.Trail, 16, 16, 0.02f,Color.red, handler ));
+       handler.addObject(new Trail((int)x,(int)y,ID.Trail, 16, 16, 0.02f,Color.red, handler ));
 
     }
 
     @Override
     public void render(Graphics g) {
       g.setColor(Color.red);
-      g.fillRect(x,y,16,16);
+      g.fillRect((int)x,(int)y,16,16);
     }
 }
